tidy up paths after reorganisation
