+++ app/routes/bannerRoutes.test.js
@@ -3,7 +3,7 @@
 const request = require('request')
 const cheerio = require('cheerio')
 
-const configPaths = require('../config/paths.json')
+const configPaths = require('../../config/paths.json')
 const PORT = configPaths.ports.test
 
 // Returns a wrapper for `request` which applies these options by default
